+++ webui/module/Application/src/Application/View/Helper/UpdateAlert.php
@@ -5,7 +5,7 @@
  * bareos-webui - Bareos Web-Frontend
  *
  * @link      https://github.com/bareos/bareos for the canonical source repository
- * @copyright Copyright (C) 2013-2023 Bareos GmbH & Co. KG (http://www.bareos.org/)
+ * @copyright Copyright (C) 2013-2024 Bareos GmbH & Co. KG (http://www.bareos.org/)
  * @license   GNU Affero General Public License (http://www.gnu.org/licenses/)
  * @author    Frank Bergkemper
  *
@@ -30,19 +30,16 @@ use Zend\View\Helper\AbstractHelper;
 
 class UpdateAlert extends AbstractHelper
 {
-    protected $value;
-    protected $result;
-
-    public function __invoke($product_updates_status = null, $dird_update_available = null)
+    public function __invoke($version_info = null)
     {
-        if ($product_updates_status === false) {
-            $this->result = '<a data-toggle="tooltip" data-placement="bottom" href="http://download.bareos.com/" target="_blank"title="Update informaton could not be retrieved"><span class="glyphicon glyphicon-exclamation-sign text-danger" aria-hidden="true"></span></a>';
-            return $this->result;
-        }
+        $result = null;
 
-        if ($dird_update_available === true) {
-            $this->result = '<a data-toggle="tooltip" data-placement="bottom" href="http://download.bareos.com/" target="_blank"title="Updates available"><span class="glyphicon glyphicon-exclamation-sign text-danger" aria-hidden="true"></span></a>';
-            return $this->result;
+        if (!$version_info) {
+            $result = '<a data-toggle="tooltip" data-placement="bottom" href="https://download.bareos.com/" target="_blank" title="Update information could not be retrieved"><span class="glyphicon glyphicon-exclamation-sign text-danger" aria-hidden="true"></span></a>';
+        } elseif ($version_info['status'] != "uptodate") {
+            $message = "Bareos Director (" . $version_info["requested_version"] . "): " . $version_info['package_update_info'];
+            $result = '<a data-toggle="tooltip" data-placement="bottom" href="https://download.bareos.com/" target="_blank" title="' . $message . '"><span class="glyphicon glyphicon-exclamation-sign text-danger" aria-hidden="true"></span></a>';
         }
+        return $result;
     }
 }